+++ src/Symfony/Tests/EventListener/SerializeListenerTest.php
@@ -92,7 +92,7 @@ class SerializeListenerTest extends TestCase
     }
 
     #[DataProvider('provideNonApiAttributes')]
-    public function testNoCallProcessor(...$attributes): void
+    public function testNoCallProcessor(array $attributes): void
     {
         $controllerResult = new \stdClass();
         $processor = $this->createMock(ProcessorInterface::class);
@@ -114,8 +114,8 @@ class SerializeListenerTest extends TestCase
     public static function provideNonApiAttributes(): array
     {
         return [
-            ['_api_respond' => false, '_api_operation_name' => 'dummy'],
-            [],
+            [['_api_respond' => false, '_api_operation_name' => 'dummy']],
+            [[]],
         ];
     }
 }